# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShiftSwap API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List available shifts with filters
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: location
          in: query
          schema:
            type: ["string", "null"]
        - name: shiftType
          in: query
          schema:
            type: ["string", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: minPay
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPay
          in: query
          schema:
            type: ["number", "null"]
        - name: mine
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftWithProfile"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Post a new shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Get a single shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithProfile"
        "404":
          description: Not found
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          description: Not found
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /shifts/{id}/requests:
    get:
      operationId: listShiftRequests
      tags: [requests]
      summary: List requests for a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftRequestWithProfile"
    post:
      operationId: createShiftRequest
      tags: [requests]
      summary: Apply to swap/cover a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftRequest"
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, clerkUserId, displayName, jobRole, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        displayName:
          type: string
        jobRole:
          type: string
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        preferredDays:
          type: array
          items:
            type: string
        preferredTimes:
          type: array
          items:
            type: string
        minHourlyRate:
          type: ["number", "null"]
        maxHourlyRate:
          type: ["number", "null"]
        isPro:
          type: boolean
        createdAt:
          type: string
...
    Shift:
      type: object
      required: [id, clerkUserId, title, role, location, shiftDate, startTime, endTime, shiftType, status, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        title:
          type: string
        role:
          type: string
        location:
          type: string
        shiftDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        hourlyRate:
          type: ["number", "null"]
        shiftType:
          type: string
          enum: [swap, cover]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, pending, filled, cancelled]
        createdAt:
          type: string
    ShiftWithProfile:
      type: object
      required: [id, clerkUserId, title, role, location, shiftDate, startTime, endTime, shiftType, status, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        title:
          type: string
        role:
          type: string
        location:
          type: string
        shiftDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        hourlyRate:
          type: ["number", "null"]
        shiftType:
          type: string
          enum: [swap, cover]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, pending, filled, cancelled]
        createdAt:
          type: string
        poster:
          $ref: "#/components/schemas/Profile"
        requestCount:
          type: integer
    ShiftInput:
      type: object
      required: [title, role, location, shiftDate, startTime, endTime, shiftType]
      properties:
        title:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        shiftDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        hourlyRate:
          type: number
        shiftType:
          type: string
          enum: [swap, cover]
        notes:
          type: string
    ShiftUpdate:
      type: object
      properties:
        title:
          type: string
        role:
          type: string
        location:
          type: string
        shiftDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        hourlyRate:
          type: number
        shiftType:
          type: string
          enum: [swap, cover]
        notes:
          type: string
        status:
          type: string
          enum: [open, pending, filled, cancelled]
    ShiftRequest:
      type: object
      required: [id, shiftId, clerkUserId, status, createdAt]
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        clerkUserId:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
    ShiftRequestInput:
      type: object
      properties:
        message:
          type: string
    ShiftRequestWithProfile:
      type: object
      required: [id, shiftId, clerkUserId, status, createdAt]
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        clerkUserId:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
        requester:
          $ref: "#/components/schemas/Profile"
...