# lib/api-zod/src/generated/types/shiftInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftInputShiftType } from './shiftInputShiftType';

export interface ShiftInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  location: string;
  shiftDate: string;
  startTime: string;
  endTime: string;
  hourlyRate?: number;
  shiftType: ShiftInputShiftType;
  notes?: string;
}