# lib/api-zod/src/generated/types/shiftRequestWithProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShiftSwap API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';
import type { ShiftRequestWithProfileStatus } from './shiftRequestWithProfileStatus';

export interface ShiftRequestWithProfile {
  id: number;
  shiftId: number;
  clerkUserId: string;
  /** @nullable */
  message?: string | null;
  status: ShiftRequestWithProfileStatus;
  createdAt: string;
  requester?: Profile;
}